Moved db population code to db.py
All the code that imports data from the csv and puts it in the database is now in db.py instead of __init__.py. The database automatically gets populated on initialization (i.e. when init_db() is called)
